* Added updated contact information for myself * Corrected SCM information * Added release numbering tied to gmaps version * Fixed broken unit test in Point * Updated Point/Size to take doubles

+++ Apis_Google_Maps/src/com/google/gwt/maps/client/base/Point.java
@@ -20,11 +20,11 @@ public class Point extends JavaScriptObject {
    * @param y
    * @return
    */
-  public final static Point newInstance(int x, int y) {
+  public final static Point newInstance(double x, double y) {
     return createJso(x, y).cast();
   }
   
-  private final static native JavaScriptObject createJso(int x, int y) /*-{
+  private final static native JavaScriptObject createJso(double x, double y) /*-{
     return new $wnd.google.maps.Point(x, y);
   }-*/;
